Implemented Device.StartTimer (#19)
This feature has been missing for a while, and it was requested on Twitter. It actually is pretty easy to implement using async/await. I added two unit tests that show that is works in general.

It remains to be seen if it will be easy to write tests against views using `Device.StartTimer`. Unless you you model your view's APIs to use `Task` and `TaskCompletionSource<T>` these views will not be very testable. If you are dealing with existing code, adding a `Task.Delay` in your tests (longer than the timer) will *work*, but it is not ideal.